Move `TestMultisigAddressGeneration` to unit tests
